requirements.txt
astroid==2.7.3
autopep8==1.5.7
colorama==0.4.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.21.2
Pillow==8.3.1
platformdirs==2.3.0
pure-python-adb==0.3.0.dev0
pycodestyle==2.7.0
pylint==2.10.2
toml==0.10.2
wrapt==1.12.1
